iio : Add cm3218 smbus ara and acpi support
On asus T100, Capella cm3218 chip is implemented as ambiant light
sensor. This chip expose an smbus ARA protocol device on standard
address 0x0c. The chip is not functional before all alerts are
acknowledged.
On asus T100, this device is enumerated on ACPI bus and the description
give two I2C connection. The first is the connection to the ARA device
and the second gives the real address of the device.

So, on device probe,If the i2c address is the ARA address and the
device is enumerated via acpi, we lookup for the real address in
the ACPI resource list and create an i2c dummy device with that address.
If the client address is already the real one, we create an i2c dummy
device for the ara function.

if an interrupt resource is given, we request a treaded interrupt to
acknowledge the smbus alert for cm3218 chip.

Signed-off-by: Marc CAPDEVILLE <m.capdeville@no-log.org>